refactor(shared): extract voting helpers into shared.voting
Council-style plugins use the same vote tally and majority rules. Share the helpers so tie and abstain semantics stay aligned.
