+++ src/owui_ext/shared/voting.py
@@ -1,0 +1,21 @@
+"""A/B voting helpers shared across owui_ext tool plugins."""
+
+from typing import List
+
+
+def compute_vote_tally(votes: List[str]) -> dict:
+    tally = {"A": 0, "B": 0, "abstain": 0}
+    for vote in votes:
+        if vote in tally:
+            tally[vote] += 1
+    return tally
+
+
+def decide_majority(tally: dict) -> str:
+    if tally.get("A", 0) > tally.get("B", 0):
+        return "A"
+    if tally.get("B", 0) > tally.get("A", 0):
+        return "B"
+    if tally.get("A", 0) == tally.get("B", 0) and tally.get("A", 0) > 0:
+        return "tie"
+    return "no_decision"
