+++ src/owui_ext/tools/magi_decision_support.py
@@ -33,6 +33,7 @@ from owui_ext.shared.prompt_utils import (
 )
 from owui_ext.shared.terminal_events import emit_terminal_tool_event
 from owui_ext.shared.tool_event_metadata import CITATION_TOOLS, TERMINAL_EVENT_TOOLS
+from owui_ext.shared.voting import compute_vote_tally, decide_majority
 
 log = logging.getLogger(__name__)
 
@@ -1053,24 +1054,6 @@ async def build_tools_dict(
             tools_dict[name] = tool_dict
 
     return tools_dict
-
-
-def compute_vote_tally(votes: List[str]) -> dict:
-    tally = {"A": 0, "B": 0, "abstain": 0}
-    for vote in votes:
-        if vote in tally:
-            tally[vote] += 1
-    return tally
-
-
-def decide_majority(tally: dict) -> str:
-    if tally.get("A", 0) > tally.get("B", 0):
-        return "A"
-    if tally.get("B", 0) > tally.get("A", 0):
-        return "B"
-    if tally.get("A", 0) == tally.get("B", 0) and tally.get("A", 0) > 0:
-        return "tie"
-    return "no_decision"
 
 
 class Tools:
